+++ packages/react/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @hello-worlds/react
 
+## 0.0.9
+
+### Patch Changes
+
+- Allow for adding r3f <material/> inside <Planet/> children
+- Updated dependencies
+  - @hello-worlds/planets@0.0.9
+
 ## 0.0.8
 
 ### Patch Changes